base/requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
dj-rest-auth==7.2.0
Django==5.2.12
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.29.0
gunicorn==25.1.0
hiredis==3.3.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pytest==8.4.2
python-dotenv==1.2.2
python3-openid==3.2.0
pytils==0.4.4
PyYAML==6.0.3
redis==7.2.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.5
